requirements.txt
psutil
python-telegram-bot
wikipedia
datetime
ping3
shodan
sentry_sdk
